Remove folder references from AngularNetApi.csproj
The code changes in the `AngularNetApi.csproj` file involve the
removal of a line that included several folder references.
Specifically, the line that was removed contained folder
inclusions for `API\Filters`, `API\Models`,
`Application\MediatR\StoreManagement`,
`Application\MediatR\AuditLog`, `Application\MediatR\OrderProcessing`,
and `Application\MediatR\Notification`.

This change suggests a restructuring of the project or the removal
of unused directories.